requirements.txt
streamlit
graphviz
schedule
tabulate
pandas
openai
python-dotenv
plotly.express
SpeechRecognition
pyaudio